+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/contentful/ContentfulBundle/compare/0.6.2-beta...HEAD)
 
+### Added
+* Exposed new features from the newer Contentful SDK
+  * Overriding URI used to communicate with the Contentful API (`uri_override`)
+
 ### Changed
 * Use version 0.7 of the Contentful SDK. **Note:** This release contains breaking changes, you can read more about them
 in the [Changelog](https://github.com/contentful/contentful.php/releases/tag/0.7.0-beta).