+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Exposed new features from the newer Contentful SDK
   * Overriding URI used to communicate with the Contentful API (`uri_override`)
+  * Setting the default locale for the client (`default_locale`)
 
 ### Changed
 * Use version 0.7 of the Contentful SDK. **Note:** This release contains breaking changes, you can read more about them